Keep the path to the resolve file object around instead of re-opening
This means we pin the dentry and inode for the file object rather than
recompute it when we open a file, so the file on disk can't be replaced.
This means, that if we trusted the inode content initially and it is
immutable via fs-verity, it won't have changed when we open it.